Update - Move DOM listeners from eventListeners into its own module
